Entities now add themselves to the world.
We encounted an issue where Box2D "locks" the world during the collision code.
This means that the asteroids cannot add entities during the onCollide method.
In order to get round this, they need to be created in the update loop, but...
entities didn't know know to build themselves and could not be passed in a set
due to the fact they required a b2Body.

I have refactored the code so that they no longer require a b2Body to be built
but only get added to the world when their init method gets called. This has
the side effect of making entities much much easier to construct (simply call
the constructor and the add method of the world).

Sorry for the long commit message but this is important.